Fixed Refactor -> Fixed things Nebula broke
