+++ src/main/java/net/apolloclient/events/CancelableEvent.java
