+++ src/main/java/net/apolloclient/events/bus/EventSubscriber.java
