+++ src/main/java/net/apolloclient/events/bus/Priority.java
