+++ src/main/java/net/apolloclient/events/impl/world/SinglePlayerJoinEvent.java
