+++ src/main/java/net/apolloclient/examples/SlothpixelExample.java
