+++ src/main/java/net/apolloclient/hud/framework/objects/GuiElement.java
