+++ src/main/java/net/apolloclient/hud/framework/objects/Rectangle.java
