+++ src/main/java/net/apolloclient/mixins/client/MixinSoundManager.java
