+++ src/main/java/net/apolloclient/mixins/entity/MixinEntityPlayer.java
@@ -1,6 +1,6 @@
 package net.apolloclient.mixins.entity;
 
-import net.apolloclient.events.impl.player.EventAttackEntity;
+import net.apolloclient.events.impl.player.AttackEntityEvent;
 import net.minecraft.client.Minecraft;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
@@ -16,10 +16,10 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(EntityPlayer.class)
 public class MixinEntityPlayer {
 
-    /** posts a {@link EventAttackEntity}. **/
+    /** posts a {@link AttackEntityEvent}. **/
     @Inject(method = "attackTargetEntityWithCurrentItem", at = @At("RETURN")) public void attackTargetEntityWithCurrentItem(Entity targetEntity, CallbackInfo ci) {
         if(targetEntity != null) {
-            EventAttackEntity event = new EventAttackEntity(Minecraft.getMinecraft().thePlayer, targetEntity);
+            AttackEntityEvent event = new AttackEntityEvent(Minecraft.getMinecraft().thePlayer, targetEntity);
         }
     }
 }
