+++ src/main/java/net/apolloclient/mixins/network/MixinNetHandlerPlayClient.java
@@ -1,6 +1,26 @@
+/*⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤
+ Copyright (C) 2020-2021 developed by Icovid and Apollo Development Team
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as published
+ by the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see https://www.gnu.org/licenses/.
+
+ Contact: Icovid#3888 @ https://discord.com
+ ⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤*/
+
 package net.apolloclient.mixins.network;
 
-import net.apolloclient.events.impl.client.ClientChatReceivedEvent;
+import net.apolloclient.events.impl.client.ActionBarEvent;
+import net.apolloclient.events.impl.client.ChatReceivedEvent;
 import net.minecraft.client.network.NetHandlerPlayClient;
 import net.minecraft.network.play.server.S02PacketChat;
 import org.spongepowered.asm.mixin.Mixin;
@@ -8,33 +28,23 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import net.apolloclient.events.impl.client.ChatReceivedEvent;
-
 /** MixinBootstrap Events for MixinNetHandlerPlayClient.class.
  * @author Nora Cos | Nora#0001
  * @since 1.0.0 **/
 @Mixin(NetHandlerPlayClient.class)
 public class MixinNetHandlerPlayClient {
 
-    /** {@link EventChatReceived} at chat packet.
+    /** Post {@link ActionBarEvent} or {@link ChatReceivedEvent} at chat packet.
      * @param callbackInfo unused
      * @param packetIn chat packet received **/
     @Inject(method = "handleChat", at = @At(value = "INVOKE", target = "Lnet/minecraft/network/PacketThreadUtil;checkThreadAndEnqueue(Lnet/minecraft/network/Packet;Lnet/minecraft/network/INetHandler;Lnet/minecraft/util/IThreadListener;)V", shift = At.Shift.AFTER), cancellable = true)
     private void onChatPacket(S02PacketChat packetIn, CallbackInfo callbackInfo) {
-        EventChatReceived event = new EventChatReceived(packetIn.getChatComponent());
+        ChatReceivedEvent event;
+        if (packetIn.getType() == 2) event = new ActionBarEvent(packetIn.getChatComponent());
+        else event = new ChatReceivedEvent(packetIn.getChatComponent());
         event.post();
         if (event.isCanceled()) {
             callbackInfo.cancel();
         }
-    }
-
-    /** {@link EventClientChatReceived} at handleChat.
-     * @author MatthewTGM | MatthewTGM#4058
-     * @param callbackInfo unused
-     * @param packetIn chat packet received. **/
-    @Inject(method = "handleChat", at = @At("HEAD"), cancellable = false)
-    private void handleChat(S02PacketChat packetIn, CallbackInfo callbackInfo) {
-        EventClientChatReceived event = new EventClientChatReceived(packetIn.getType(), packetIn.getChatComponent());
-        event.post();
     }
 }
