+++ src/main/java/net/apolloclient/modules/Category.java
