+++ src/main/java/net/apolloclient/modules/ModuleManager.java
