+++ src/main/java/net/apolloclient/modules/impl/util/DiscordRPModule.java
