+++ src/main/java/net/apolloclient/modules/impl/util/WindowedFullscreenModule.java
