+++ src/main/java/net/apolloclient/settings/Setting.java
